fetch anrop för att hämta alla requests korrekt nu
